* Fix a warning * Update ChangeLog

+++ ext/standard/fsock.c
@@ -17,7 +17,7 @@
    +----------------------------------------------------------------------+
 */
 
-/* $Id: fsock.c,v 1.15 1999/07/21 16:12:13 sas Exp $ */
+/* $Id: fsock.c,v 1.16 1999/07/24 14:36:48 zeev Exp $ */
 
 /* Synced with php3 revision 1.121 1999-06-18 [ssb] */
 /* Synced with php3 revision 1.133 1999-07-21 [sas] */
@@ -248,7 +248,7 @@ static void _php3_fsockopen(INTERNAL_FUNCTION_PARAMETERS, int persistent) {
 	switch(arg_count) {
 		case 5:
 			convert_to_double(args[4]);
-			conv = args[4]->value.dval * 1000000.0;
+			conv = (unsigned long) args[4]->value.dval * 1000000.0;
 			timeout.tv_sec = conv / 1000000;
 			timeout.tv_usec = conv % 1000000;
 			/* fall-through */
